Fixed #22/#73: made everything except ML install optional, dont update os by default

+++ opt/vagrant/setup-git.sh
@@ -1,36 +1,47 @@
 #! /bin/sh
 echo "running $0 $@"
 
-# Install git
-yum -y install git
+# Defaults
+install_git=true
+install_git_project=true
 
-# create a local git repo
-mkdir -p /space/projects/$1.git
+# Load the normalized project properties.
+source /tmp/$1.project.properties
 
-# initialize the git repo
-cd /space/projects/$1.git
-if [ ! -f HEAD ]; then
-  git --bare init
-fi
-if [ ! -f hooks/post-receive ]; then
-  printf "#!/bin/sh\nin=\$(cat)\nbranch=\${in##*/}\nGIT_WORK_TREE=/space/projects/$1.live git checkout -f \$branch\nchown -R :sshuser /space/projects/$1.live\nchmod -R g+rw /space/projects/$1.live\n" > hooks/post-receive
-  chmod 755 hooks/post-receive
+if [ install_git == true ]; then
+  # Install git
+  yum -y install git
 fi
 
-# create a project deploy dir
-mkdir -p /space/projects/$1.live
+if [ install_git_project == true ]; then
+  # create a local git repo
+  mkdir -p /space/projects/$1.git
 
-# change permissions to allow remote deployment
-cd /space/projects
-if [ -d /vagrant ]; then
-  chown -R $1:sshuser $1.git
-  chown -R $1:sshuser $1.live
-else
-  # creation of users is limited on demo servers
-  chown -R $USER:sshuser $1.git
-  chown -R $USER:sshuser $1.live
+  # initialize the git repo
+  cd /space/projects/$1.git
+  if [ ! -f HEAD ]; then
+    git --bare init
+  fi
+  if [ ! -f hooks/post-receive ]; then
+    printf "#!/bin/sh\nin=\$(cat)\nbranch=\${in##*/}\nGIT_WORK_TREE=/space/projects/$1.live git checkout -f \$branch\nchown -R :sshuser /space/projects/$1.live\nchmod -R g+rw /space/projects/$1.live\n" > hooks/post-receive
+    chmod 755 hooks/post-receive
+  fi
+
+  # create a project deploy dir
+  mkdir -p /space/projects/$1.live
+
+  # change permissions to allow remote deployment
+  cd /space/projects
+  if [ -d /vagrant ]; then
+    chown -R $1:sshuser $1.git
+    chown -R $1:sshuser $1.live
+  else
+    # creation of users is limited on demo servers
+    chown -R $USER:sshuser $1.git
+    chown -R $USER:sshuser $1.live
+  fi
+  chmod -R g+rw $1.git
+  chmod -R g+rw $1.live
+  chmod g+s $1.git
+  chmod g+s $1.live
 fi
-chmod -R g+rw $1.git
-chmod -R g+rw $1.live
-chmod g+s $1.git
-chmod g+s $1.live
